+++ app/controllers/staff/activities_controller.rb
@@ -2,8 +2,11 @@
 
 class Staff::ActivitiesController < Staff::BaseController
   include Secured
+  include Activities::Breadcrumbed
+
   after_action :verify_authorized, except: [:index, :historic]
   after_action :skip_policy_scope, only: [:index, :historic]
+  skip_before_action :clear_breadcrumb_context, only: [:show]
 
   def index
     @organisation = Organisation.find(organisation_id)
@@ -17,7 +20,12 @@ class Staff::ActivitiesController < Staff::BaseController
         organisation: @organisation,
         scope: :current
       ).call
-      add_breadcrumb t("page_content.breadcrumbs.current_index"), organisation_activities_path(@organisation)
+
+      if current_user.service_owner?
+        add_breadcrumb t("page_content.breadcrumbs.organisation_current_index", org_name: @organisation.name), organisation_activities_path(@organisation)
+      else
+        add_breadcrumb t("page_content.breadcrumbs.current_index"), organisation_activities_path(@organisation)
+      end
     end
   end
 
@@ -25,17 +33,23 @@ class Staff::ActivitiesController < Staff::BaseController
     @activity = Activity.find(id)
     authorize @activity
 
-    @activities = @activity.child_activities.order("created_at ASC").map { |activity| ActivityPresenter.new(activity) }
-
-    @transactions = policy_scope(Transaction).where(parent_activity: @activity).order("date DESC")
-    @budgets = policy_scope(Budget).where(parent_activity: @activity).order("financial_year DESC")
-    @forecasts = policy_scope(@activity.latest_forecasts)
-
     respond_to do |format|
       format.html do
-        redirect_to organisation_activity_financials_path(@activity.organisation, @activity)
+        prepare_default_activity_trail(@activity)
+
+        tab = Activity::Tab.new(
+          activity: @activity,
+          current_user: current_user,
+          tab_name: current_tab
+        )
+        render template: tab.template, locals: tab.locals
       end
       format.xml do |_format|
+        @activities = @activity.child_activities.order("created_at ASC").map { |activity| ActivityPresenter.new(activity) }
+
+        @transactions = policy_scope(Transaction).where(parent_activity: @activity).order("date DESC")
+        @budgets = policy_scope(Budget).where(parent_activity: @activity).order("financial_year DESC")
+        @forecasts = policy_scope(@activity.latest_forecasts)
         @reporting_organisation = Organisation.service_owner
 
         response.headers["Content-Disposition"] = "attachment; filename=\"#{@activity.transparency_identifier}.xml\""
@@ -54,14 +68,19 @@ class Staff::ActivitiesController < Staff::BaseController
         organisation: @organisation,
         scope: :historic
       ).call
-      add_breadcrumb t("page_content.breadcrumbs.historic_index"), historic_organisation_activities_path(@organisation)
+
+      if current_user.service_owner?
+        add_breadcrumb t("page_content.breadcrumbs.organisation_historic_index", org_name: @organisation.name), historic_organisation_activities_path(@organisation)
+      else
+        add_breadcrumb t("page_content.breadcrumbs.historic_index"), historic_organisation_activities_path(@organisation)
+      end
     end
   end
 
   private
 
   def id
-    params[:id]
+    params[:id] || params[:activity_id]
   end
 
   def organisation_id
@@ -79,5 +98,9 @@ class Staff::ActivitiesController < Staff::BaseController
 
   def fund_id
     params[:fund_id]
+  end
+
+  def current_tab
+    params[:tab] || "financials"
   end
 end