+++ app/services/activity/tab.rb
@@ -1,0 +1,80 @@
+class Activity
+  class Tab
+    include Pundit
+
+    VALID_TAB_NAMES = [
+      "financials",
+      "other_funding",
+      "details",
+      "children",
+      "comments",
+      "transfers",
+      "historical_events",
+    ].freeze
+
+    def initialize(activity:, current_user:, tab_name: "financials")
+      raise ActionController::RoutingError.new("invalid tab") unless VALID_TAB_NAMES.include?(tab_name)
+
+      @tab_name = tab_name
+      @activity = ActivityPresenter.new(activity)
+      @current_user = current_user
+
+      send(tab_name)
+    end
+
+    def locals
+      instance_variables.map { |attribute|
+        [attribute, instance_variable_get(attribute)]
+      }.to_h
+    end
+
+    def template
+      "staff/activities/#{tab_name}"
+    end
+
+    private
+
+    attr_reader :current_user, :tab_name
+
+    def financials
+      @transactions = policy_scope(Transaction).where(parent_activity: @activity).order("date DESC")
+      @budgets = policy_scope(Budget).where(parent_activity: @activity).order("financial_year DESC")
+      @forecasts = policy_scope(@activity.latest_forecasts)
+      @refunds = policy_scope(Refund).where(parent_activity: @activity).order("financial_year DESC")
+
+      @transaction_presenters = @transactions.includes(:parent_activity).map { |transaction| TransactionPresenter.new(transaction) }
+      @budget_presenters = @budgets.includes(:parent_activity, :providing_organisation).map { |budget| BudgetPresenter.new(budget) }
+      @forecast_presenters = @forecasts.map { |forecast| ForecastPresenter.new(forecast) }
+      @refund_presenters = @refunds.map { |forecast| RefundPresenter.new(forecast) }
+
+      @implementing_organisation_presenters = @activity.implementing_organisations.map { |implementing_organisation| ImplementingOrganisationPresenter.new(implementing_organisation) }
+    end
+
+    def other_funding
+      @matched_efforts = @activity.matched_efforts.map { |e| MatchedEffortPresenter.new(e) }
+      @external_incomes = @activity.external_incomes.map { |e| ExternalIncomePresenter.new(e) }
+    end
+
+    def details
+      @activities = @activity.child_activities.order("created_at ASC").map { |activity| ActivityPresenter.new(activity) }
+      @implementing_organisation_presenters = @activity.implementing_organisations.map { |implementing_organisation| ImplementingOrganisationPresenter.new(implementing_organisation) }
+    end
+
+    def children
+      @activities = @activity.child_activities.includes([:organisation, :parent]).order("created_at ASC").map { |activity| ActivityPresenter.new(activity) }
+    end
+
+    def comments
+      @comments = Comment.where(activity_id: @activity.id).includes(:report)
+    end
+
+    def transfers
+      @outgoing_transfers = policy_scope(OutgoingTransfer.where(source: @activity)).map { |transfer| TransferPresenter.new(transfer) }
+      @incoming_transfers = policy_scope(IncomingTransfer.where(destination: @activity)).map { |transfer| TransferPresenter.new(transfer) }
+    end
+
+    def historical_events
+      @historical_events = Activity::HistoricalEventsGrouper.new(activity: @activity).call
+    end
+  end
+end