+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2021_07_27_163803) do
+ActiveRecord::Schema.define(version: 2021_08_11_093658) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -243,6 +243,19 @@ ActiveRecord::Schema.define(version: 2021_07_27_163803) do
     t.index ["source_id"], name: "index_outgoing_transfers_on_source_id"
   end
 
+  create_table "refunds", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
+    t.uuid "parent_activity_id"
+    t.uuid "report_id"
+    t.integer "financial_year"
+    t.integer "financial_quarter"
+    t.decimal "value", precision: 13, scale: 2
+    t.text "comment"
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
+    t.index ["parent_activity_id"], name: "index_refunds_on_parent_activity_id"
+    t.index ["report_id"], name: "index_refunds_on_report_id"
+  end
+
   create_table "reports", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "state", default: "inactive", null: false
     t.string "description"
@@ -301,11 +314,11 @@ ActiveRecord::Schema.define(version: 2021_07_27_163803) do
   add_foreign_key "activities", "organisations", on_delete: :restrict
   add_foreign_key "budgets", "activities", column: "parent_activity_id", on_delete: :cascade
   add_foreign_key "budgets", "organisations", column: "providing_organisation_id"
+  add_foreign_key "forecasts", "activities", column: "parent_activity_id", on_delete: :cascade
   add_foreign_key "incoming_transfers", "activities", column: "destination_id", on_delete: :restrict
   add_foreign_key "incoming_transfers", "activities", column: "source_id", on_delete: :restrict
   add_foreign_key "outgoing_transfers", "activities", column: "destination_id", on_delete: :restrict
   add_foreign_key "outgoing_transfers", "activities", column: "source_id", on_delete: :restrict
-  add_foreign_key "forecasts", "activities", column: "parent_activity_id", on_delete: :cascade
   add_foreign_key "transactions", "activities", column: "parent_activity_id", on_delete: :cascade
   add_foreign_key "users", "organisations", on_delete: :restrict
 end