+++ spec/controllers/staff/activities_controller_spec.rb
@@ -116,4 +116,38 @@ RSpec.describe Staff::ActivitiesController do
   describe "#historic" do
     include_examples "fetches activities", {route: :historic, scope: :historic}
   end
+
+  describe "#show" do
+    context "when viewing historical events" do
+      let(:user) { create(:delivery_partner_user, organisation: organisation) }
+      let(:organisation) { create(:delivery_partner_organisation) }
+
+      before do
+        allow(controller).to receive(:current_user).and_return(user)
+        allow(controller).to receive(:logged_in_using_omniauth?).and_return(true)
+        allow(controller).to receive(:prepare_default_activity_trail)
+      end
+
+      let(:activity) { build_stubbed(:project_activity, organisation: organisation) }
+      let(:grouper) { instance_double(Activity::HistoricalEventsGrouper, call: double) }
+
+      before do
+        allow(Activity).to receive(:find).and_return(activity)
+        allow(ActivityPresenter).to receive(:new).and_return(activity)
+        policy = instance_double(ActivityPolicy, show?: true)
+        allow(ActivityPolicy).to receive(:new).and_return(policy)
+        allow(Activity::HistoricalEventsGrouper).to receive(:new).and_return(grouper)
+      end
+
+      it "asks the HistoricalEventsGrouper to prepare the 'Change history'" do
+        get :show, params: {activity_id: "abc123", organisation_id: organisation.id, tab: "historical_events"}
+
+        expect(Activity::HistoricalEventsGrouper)
+          .to have_received(:new)
+          .with(activity: activity)
+
+        expect(grouper).to have_received(:call)
+      end
+    end
+  end
 end